Add a cancel command for teaching assistants.
The cancel command for teaching assistants will cancel the waiting
status that is set if the assistant has typed !next when the queue was
empty. In other words, the assistant will have to type !next to get the
next student again, and will not be notified when a new student requests
help.